Develop fatiyya

+++ src/main/java/com/e_kampoeng/model/WilayahRTModel.java
@@ -3,7 +3,7 @@ package com.e_kampoeng.model;
 import javax.persistence.*;
 
 @Entity
-@Table(name="wilayah_rt")
+@Table(name = "wilayah_rt")
 public class WilayahRTModel {
 
     @Id
@@ -11,7 +11,7 @@ public class WilayahRTModel {
     private Long id;
 
     @Column(name = "nomor_rt")
-    private Long nomor_rt;
+    private Integer nomor_rt;
 
     public WilayahRTModel() {
     }
@@ -24,11 +24,11 @@ public class WilayahRTModel {
         this.id = id;
     }
 
-    public Long getNomor_rt() {
+    public Integer getNomor_rt() {
         return nomor_rt;
     }
 
-    public void setNomor_rt(Long nomor_rt) {
+    public void setNomor_rt(Integer nomor_rt) {
         this.nomor_rt = nomor_rt;
     }
 }
